Fixing time delta issue. Adding daemon scripts.
Rather than installing python 2.7 (not supported on Ubuntu 10)
	implemented the functionality of the total_second() function.

Added to scripts for administration.

zpd.sh gets Kerberos tickets using the keytab file, then starts
zwgc. This is used by the init script.

django_admin.sh is a script to easily start the django admin. Note that
it binds to 0.0.0.0:8000 so it is available from any interface.