Fixing classes being added twice.
addZephyrClass was pushing to api.classes, but did not have
all the fields such as instanceDict and also did not add the
subscription to the classDict.

Deleted this code from addZephyrClass, the addSubscription takes care of
creating the class correctly and adding it to api.classDict.

The class sidebar does not display correctly, but the internal state is
now correct for adding classes.

Also handling if Add Class is cancelled, cleanly returns.