Turning off DEBUG
django was caching sql queries causing memory to grow (i think this is
the cause for the memory leak)

Turning off DEBUG to see if this fixes things. Alternatively
django.db.reset_queries() in loadZephyrs if we still want DEBUG on.

Since debug is on, I am having django email me. Assuming it doesn't spam
me with emails, I'll change the email to zephyrplus.